+++ src/main/scala/org/fs/mael/core/backend/BackendDownloader.scala
@@ -25,6 +25,21 @@ abstract class BackendDownloader(protected val backendId: String) {
 
   def startInner(de: DownloadEntry, timeoutMs: Int): Unit
 
+  /**
+   * Restart downloading the given entry with the given timeout
+   * @param timeoutMs operations timeout in millis, 0 stands for none
+   */
+  def restart(de: DownloadEntry, timeoutMs: Int): Unit = {
+    require(de.status != Status.Running, "Can't restart an ongoing download!")
+    de.fileOption foreach (_.delete())
+    addLogAndFire(de, LogEntry.info("Restarting download from the beginning"))
+    de.sections.clear()
+    de.supportsResumingOption = None
+    eventMgr.fireDetailsChanged(de)
+    changeStatusAndFire(de, Status.Stopped)
+    start(de, timeoutMs)
+  }
+
   def stop(de: DownloadEntry): Unit = de.status match {
     case s if s.canBeStopped => stopInner(de)
     case _                   => // NOOP