+++ src/main/scala/org/fs/mael/core/event/Events.scala
@@ -2,7 +2,6 @@ package org.fs.mael.core.event
 
 import org.fs.mael.core.Status
 import org.fs.mael.core.entry.DownloadEntry
-import org.fs.mael.core.entry.DownloadEntryView
 import org.fs.mael.core.entry.LogEntry
 
 sealed trait PriorityEvent {
@@ -46,7 +45,7 @@ object Events {
    *
    * Should be fired by download list manager.
    */
-  case class Added(de: DownloadEntryView) extends PriorityEventImpl(100) with EventForUi {
+  case class Added(de: DownloadEntry) extends PriorityEventImpl(100) with EventForUi {
     override def msg = "Added " + de.uri
   }
 
@@ -55,7 +54,7 @@ object Events {
    *
    * Should be fired by download list manager.
    */
-  case class Removed(de: DownloadEntryView) extends PriorityEventImpl(100) with EventForUi {
+  case class Removed(de: DownloadEntry) extends PriorityEventImpl(100) with EventForUi {
     override def msg = "Removed " + de.uri
   }
 
@@ -68,7 +67,7 @@ object Events {
    *
    * Should be fired by backend.
    */
-  case class StatusChanged(de: DownloadEntryView, prevStatus: Status) extends PriorityEventImpl(100) with EventForUi {
+  case class StatusChanged(de: DownloadEntry, prevStatus: Status) extends PriorityEventImpl(100) with EventForUi {
     override def msg = "Status of " + de.uri + " changed from " + prevStatus + " to " + de.status
   }
 
@@ -77,7 +76,7 @@ object Events {
    *
    * Should be fired by backend.
    */
-  case class DetailsChanged(de: DownloadEntryView) extends PriorityEventImpl(50) with EventForUi {
+  case class DetailsChanged(de: DownloadEntry) extends PriorityEventImpl(50) with EventForUi {
     override def msg = "Details changed for " + de.uri
   }
 
@@ -86,7 +85,7 @@ object Events {
    *
    * Should be fired by backend.
    */
-  case class Logged(de: DownloadEntryView, entry: LogEntry) extends PriorityEventImpl(20) with EventForUi {
+  case class Logged(de: DownloadEntry, entry: LogEntry) extends PriorityEventImpl(20) with EventForUi {
     override def msg = "Log entry added for " + de.uri
   }
 
@@ -97,7 +96,7 @@ object Events {
    *
    * (Note that these events will be fired much more often than UI would wish to process.)
    */
-  case class Progress(de: DownloadEntryView) extends PriorityEventImpl(Int.MinValue) with EventForUi {
+  case class Progress(de: DownloadEntry) extends PriorityEventImpl(Int.MinValue) with EventForUi {
     override def msg = "Progress for " + de.uri
   }