+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### SNAPSHOT
+* Added ability to restart download from the beginning
 * Added tray icon and minimize to tray setting
 * (Internal) Implemented backend-specific settings framework (unused yet)
   * (Internal) Implemented global defaults serving as template for local per-download settings